refactor: Preferences 속성 종류를 클래스로 분리
