+++ core/preferences/src/main/java/com/practice/preferences/PreferencesRepository.kt
@@ -1,5 +1,8 @@
 package com.practice.preferences
 
+import com.practice.preferences.preferences.ScreenMode
+import com.practice.preferences.preferences.ThemeMode
+import com.practice.preferences.preferences.UiMode
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
