+++ core/preferences/src/main/java/com/practice/preferences/preferences/ThemeMode.kt
@@ -1,0 +1,7 @@
+package com.practice.preferences.preferences
+
+enum class ThemeMode {
+    Light,
+    Dark,
+    SystemDefault,
+}
