+++ core/preferences/src/main/java/com/practice/preferences/preferences/UiMode.kt
@@ -1,0 +1,6 @@
+package com.practice.preferences.preferences
+
+enum class UiMode {
+    Graphic,
+    ScreenReader,
+}
