+++ feature/main/src/main/java/com/practice/main/VerticalMainScreen.kt
@@ -33,7 +33,7 @@ import com.practice.main.state.MainUiState
 import com.practice.main.state.MealUiState
 import com.practice.main.state.MemoUiState
 import com.practice.main.state.ScheduleUiState
-import com.practice.preferences.ScreenMode
+import com.practice.preferences.preferences.ScreenMode
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
