+++ feature/main/src/main/java/com/practice/main/state/ScreenModeIcon.kt
@@ -4,7 +4,7 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Event
 import androidx.compose.material.icons.filled.List
 import androidx.compose.ui.graphics.vector.ImageVector
-import com.practice.preferences.ScreenMode
+import com.practice.preferences.preferences.ScreenMode
 
 data class ScreenModeIcon(
     val screenMode: ScreenMode,
